Updating gitignore

+++ Inky.py
@@ -1,11 +1,19 @@
 ################################################################################
-#----------------BoPyWiki---by-Brian-Evans---------------------------------------#
+#----------------InkyWiki---by-Brian-Evans---------------------------------------#
 ################################################################################
 
 from bottle import route, run, template, static_file, post, request, get
 import os.path, os
+import re
+import inkyconfig as cfg
 
 #####---------------------Main-page-view-functions-------------------------#####
+@route('/')
+def home():
+    pagename = 'main'
+    with open('./content/main', 'r') as textfile:
+        bodytext = textfile.read()
+    return template('core', pagename=pagename, bodytext=bodytext, options=cfg.inky)
 
 @route('/wiki/<pagename>')
 def build(pagename='main'):
@@ -13,9 +21,11 @@ def build(pagename='main'):
     if not os.path.isfile('./content/' + str(pagename)):
         return template('editcreate', pagename=pagename)
     else:
+        if not re.search(cfg.inky['name_re'],pagename):
+            pagename = 'main'
         with open('./content/' + pagename, 'r') as textfile:
             bodytext = textfile.read()
-        return template('core', pagename=pagename, bodytext=bodytext)
+        return template('core', pagename=pagename, bodytext=bodytext, options=cfg.inky)
 
 
 @route('/edit/<pagename>', method='POST')
@@ -25,6 +35,7 @@ def createedit(pagename):
     pagedata = stripHTML(pagedata)
     pagedata = alinkbuild(pagedata)
     pagedata = swapScript(pagedata,'ToHTML')
+    pagename = pagename.lower()
     fo = open('./content/' + pagename, 'w')
     fo.write(str.strip(pagedata))
     fo.close()
@@ -45,10 +56,15 @@ def editExisting(pagename):
 @route('/rename/<pagename>', method='POST')
 def renameExisting(pagename):
     #Renames an existing page and is fed from a function titled 'renameEntry'
-    newName = request.forms.get('newpagename')
-    newNameForURL = newName.replace(' ','_')
-    os.rename('./content/' + pagename, './content/' + newNameForURL)
-    return template('renamecomplete', pagename=pagename, newName=newName, newNameForURL=newNameForURL)
+    newName = request.forms.get('newpagename').lower()
+    if re.search(cfg.inky['name_re'],newName):
+        newNameForURL = newName.replace(' ','_')
+        os.rename('./content/' + pagename, './content/' + newNameForURL)
+        return template('renamecomplete', pagename=pagename, newName=newName, newNameForURL=newNameForURL)
+    else:
+        with open('./content/' + pagename, 'r') as textfile:
+            bodytext = textfile.read()
+        return template('core', pagename=pagename, bodytext=bodytext)
 
 
 @route('/rename/existing/<pagename>')
@@ -59,12 +75,12 @@ def renameEntry(pagename):
 
 @route('/delete/<pagename>')
 def deleteArticle(pagename):
-    if os.path.isfile('./content/' + pagename):
+    if os.path.isfile('./content/' + pagename) and re.search(cfg.inky['name_re'],pagename):
         os.remove('./content/' + pagename)
         deletestatus = 'was successful.'
         return template('deletecomplete', pagename=pagename, deletestatus=deletestatus)
     else:
-        deletestatus = 'failed. The article did not exist.'
+        deletestatus = 'failed. The article did not exist or was not a valid target for deletion.'
         return template('deletecomplete', pagename=pagename, deletestatus=deletestatus)
 
 
@@ -73,26 +89,41 @@ def servestyle(css):
     #This serves the CSS to the templates
     return static_file(css,root='./styles/')
 
+@route('/js/<js>')
+def servestyle(js):
+    #This serves the js to the templates
+    return static_file(js,root='./js/')
+
 
 @route('/wiki/search', method='POST')
 def search():
+    pagename = request.forms.get('pagename').lower()
+    if re.search(cfg.inky['name_re'],pagename):
+        pagename = pagename.replace(' ','_')
+        results = findPages(pagename)
+        found = False
 
-    pagename = request.forms.get('pagename')
-    pagename = pagename.replace(' ','_')
-    #Serves either an edit dialog or an existing page
-    if not os.path.isfile('./content/' + str(pagename)):
-        return template('editcreate', pagename=pagename)
+        for item in results:
+            if item == pagename:
+                with open('./content/' + pagename, 'r') as textfile:
+                    bodytext = textfile.read()
+                return template('core', pagename=pagename, bodytext=bodytext, options=cfg.inky)
+            else:
+                found = False
+        return template('search-results', results=results, pagename=pagename, found=found, options=cfg.inky)
     else:
+        pagename = 'main'
         with open('./content/' + pagename, 'r') as textfile:
             bodytext = textfile.read()
         return template('core', pagename=pagename, bodytext=bodytext)
 
-
-
-
-
+#
+##
+###
+####
 #####-------------Functions-for-editor(s)/String-Functions-----------------#####
-
+##--~~
+##--~~
 def alinkbuild(string):
 #Will build hyperlinks from WikiCode entered into the edit textareabox
 #At present this allows for properly nested wikiCode within '[F: ... :F]' tags
@@ -108,48 +139,44 @@ def alinkbuild(string):
                 linkname = linkname.replace(' ','_')
             else:
                 linkname = linktext.replace(' ','_')
-            string = string.replace('[F: ' + linktext + ' :F]', '<a href="/wiki/' + linkname + '">' + linktext + '</a>')
+
+            #deal with aliased links
+            original = linktext
+            if linkname.find('=') >= 1:
+                split = linkname.find('=')
+                linkname = linkname[:split]
+                linktext = linktext[split+1:]
+            string = string.replace('[F: ' + original + ' :F]', '<a href="/wiki/' + linkname + '">' + linktext + '</a>')
         return string
     else:
         return string
-
+##--~~
+##--~~
 def alinkunbuild(string):
 #Deconstructs hyperlinks back into Wikicode for the edit textarea box
     instances = string.lower()
     instances = instances.count('<a href=')
-    linklist = []
     marker = 0
     if instances > 0:
         for x in range(instances):
-            start = string.lower().find('<a href=',marker)
-            end = string.find('>',start + 1) + 1
-            linklist.append(string[start:end])
+            start = string.lower().find('<a href="/wiki/',marker)
+            end = string.find('">',start + 1) + 1
+            aliasStart = end + 1
+            aliasEnd = string.find('<',end)
+            if string[start+15:end-1] == string[aliasStart:aliasEnd]:
+                string = string.replace(string[start:end+1],'[F: ')
+            else:
+                #deal with aliased links
+                replacee = string[start:end+1]
+                replacement = string[start+15:end-1]
+                string = string.replace(replacee,'[F: ' + replacement + '=')
             marker = end + 1
-        print set(linklist)
-        for y in set(linklist):
-            string = string.replace(y,'[F: ')
         string = string.replace('</a>',' :F]')
         return string
     else:
         return string
-
-#old version
-"""def alinkunbuild(string):
-#Deconstructs hyperlinks back into Wikicode for the edit textarea box
-    instances = string.lower()
-    instances = instances.count('<a href=')
-    if instances > 0:
-        for x in range(instances):
-            start = string.lower().find('<a href=')
-            end = string.find('>',start + 1) + 1
-            linktext = string[start:end]
-            print linktext
-            string = string.replace(linktext,'[F: ')
-            string = string.replace('</a>',' :F]')
-        return string
-    else:
-        return string"""
-
+##--~~
+##--~~
 def stripHTML(string):
 #Will strip all html tags from a string
     instances = string.count('<')
@@ -162,7 +189,8 @@ def stripHTML(string):
         return string
     else:
         return string
-
+##--~~
+##--~~
 def swapScript(string,direction='FromHTML'):
     #Direction takes: 'FromHTML' and any other value (to go ToHTML)
     #Swaps html for wikiCode and vice versa
@@ -185,8 +213,17 @@ def swapScript(string,direction='FromHTML'):
         else:
             string = string.replace(x[1],x[0])
     return string
-
-
+##--~~
+##--~~
+def findPages(string):
+    #Used in serach
+    matches = []
+    for item in os.listdir('./content/'):
+        if re.search(r''+string,item,re.I):
+            matches.append(item)
+    return matches
+##--~~
+##--~~
 #####---------------------------Run-the-server-----------------------------#####
 if __name__ == '__main__':
     run(host='localhost', port=8080)
